requirements.txt
qrcode
Pillow
pyshorteners